refactor: 흩어져 있던 상수 한 파일에 모으는 작업
