+++ src/view/InputView.js
@@ -1,17 +1,13 @@
 const Console = require('../utils/Console');
-
-const REQUEST_MSG = {
-  carNames: '경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).\n',
-  raceRound: '시도할 회수는 몇회인가요?\n',
-};
+const { REQUEST_MESSAGE } = require('../constants/Constant');
 
 const InputView = {
   readCarNames(callback) {
-    Console.question(REQUEST_MSG.carNames, callback);
+    Console.question(REQUEST_MESSAGE.carNames, callback);
   },
 
   readRaceRound(callback) {
-    Console.question(REQUEST_MSG.raceRound, callback);
+    Console.question(REQUEST_MESSAGE.raceRound, callback);
   },
 };
